# 5.UNG_DUNG/5.1.backend_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HitRadar Pro API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/model-info": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Model Info",
        "description": "Return model metadata and test-set metrics.\n\nRaises 503 if the model is not loaded.",
        "operationId": "model_info_model_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfoResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/features": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Features",
        "description": "Return the 18 canonical input fields and 31 selected feature names.\n\nRaises 503 if the model is not loaded.",
        "operationId": "features_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeaturesResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "issue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "example": {
          "details": [],
          "error_code": "INVALID_FEATURE",
          "message": "Feature is not a valid canonical input field.",
          "timestamp": "2026-08-05T10:00:00Z"
        }
      },
...
      "FeaturesResponse": {
        "properties": {
          "canonical_fields": {
            "items": {
              "$ref": "#/components/schemas/FieldDescriptor"
            },
            "type": "array",
            "title": "Canonical Fields"
          },
          "selected_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Features"
          },
          "total_input_fields": {
            "type": "integer",
            "title": "Total Input Fields"
          },
          "total_selected_features": {
            "type": "integer",
            "title": "Total Selected Features"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "canonical_fields",
          "selected_features",
          "total_input_fields",
          "total_selected_features"
        ],
        "title": "FeaturesResponse"
      },
      "FieldDescriptor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          },
          "allowed_categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Categories"
          },
          "default_policy": {
            "type": "string",
            "title": "Default Policy",
            "default": "PIPELINE_IMPUTE"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "position",
          "data_type",
          "required"
        ],
        "title": "FieldDescriptor"
      },
...
      "Metrics": {
        "properties": {
          "MAE": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae"
          },
          "RMSE": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rmse"
          },
          "R2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R2"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Metrics"
      },
      "ModelInfoResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "model_family": {
            "type": "string",
            "title": "Model Family"
          },
          "package_version": {
            "type": "string",
            "title": "Package Version"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          },
          "feature_set": {
            "type": "string",
            "title": "Feature Set"
          },
          "training_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Date"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_id",
          "model_version",
          "model_family",
          "package_version",
          "data_version",
          "feature_set"
        ],
        "title": "ModelInfoResponse"
      },
...
    }
  }
}